Cleanup use of ampersands in internally generated URLs
This should fix a number of bugs related to invalid links (created by
confusion about whether to use & or &amp; in internal URLs). We should
only apply htmlspecialchars() or similar at output time, assuming we
want HTML output from Mantis. Internally, it is best to just use & to
separate arguments in GET HTTP URLs. There are still a few ugly bits of
code that haven't been updated because to do so would require extensive
rewriting in some sections. And in time, the print_ and html_ API is
probably going to be deprecated by use of a templating engine so there
is little point rewriting something that'll only get overhauled later.

This commit is the backport to the 1.2.x branch because it fixes a few
bugs here and there.

Conflicts:
	main_page.php