Fix #11229: Fix tagging XSS scripting vulnerabilities
Tag names and descriptions were not properly sanitised before being
written to HTML output. This meant that it was possible for users to
create tags containing Javascript that is executed on every load of
view_all_bug_page (and elsewhere) for all users.

Thanks to Michel Arboi from Tenable Network Security (Nessus) for
reporting this issue.

This is a backport of 70b5022f556c9b9b6b0cd661e3357767a3b178c5