Fix #11243: XSS on view_all_bug_page.php due to bad sanitising defaults
Columns on view_all_bug_page.php are not sanitised by default when there
is no special function defined for formatting and printing the column
value. This leads to a problem where a column such as 'version' can
introduce an XSS flaw when a malicious user has the ability to create
their own versions containing Javascript.

For columns with existing printing/formatting functions, these have been
improved with the use of string sanitisation where applicable.