Implemented multi-part plugin dependency checks
Plugins can now declare multi-part version dependencies, separating
parts with commas.  This allows plugins to declare both a minimum and
maximum version dependency at the same time, or potentially multiple
minimum and maximum parts.

Also, for consistency with other plugin dependency systems (such as
Mozilla's), the plugin manager has moved from a less-than to a
less-than-or-equal dependency measure, to easily specify the highest
known-compatible version in a dependency declaration.  As part of this,
dependencies can use either the existing "< XYZ" or a more-correct
"<= XYZ" string, and both will be interpreted as lte.