Fix #11261: XSS in error output as MantisCoreFormatting isn't loaded
print_project_menu_bar() is called when an error occurs in MantisBT (to
produce the HTML output for the error page). At this point of time,
MantisCoreFormatting may not be loaded by MantisBT and therefore the
string_display_* sanitisation functions won't be executed. Thus we must
force the use of a the string_html_specialchars() function to ensure
that these strings are safely sanitised even when MantisCoreFormatting
isn't loaded (yet).