Update libraries to match those of 1.3.x branch
This syncs the libraries used in the 1.2.x branch with those used in the
1.3.x branch (as of the current time). Essentially this involves the
following upgrades:

ADOdb: v5.08a to v5.10
ez Components: v2008.2.3 to v2009.1.2
PHPMailer: v5.02 to v5.1

Changes from the following commits in the 1.3.x branch are rolled up
into this patch:
f3fdde75b8548fda404a52c21f09ebf57d9c45ee
73a2b8fc7c5898ce3561bbb926c050a0218b4cab

By syncing to the latest available versions of these libraries, we're in
a much better position to respond quickly and easily to upstream
releases that may contain important security or bug fixes. We don't have
to backport patches manually if we stay up to date with upstream.