Reorder global_settings configuration option, expand _file$ rule
Sort the global_settings configuration option alphabetically, one rule
per line. This makes it easier to read and update in the future.

Additionally, the old _file$ rule has been updated to cover more
configuration settings, specifically those dealing with attachments and
uploads. All of these should be system-wide settings with the exception
of the file upload thresholds (this has been catered for in this patch).