Improve support for multiple resolved and closed statuses (bug_update)
It's possible to have multiple resolved or closed statuses. Thus when
updating a bug, we need to check not if the new status meets the
resolved or closed status threshold, but whether the new status is
within a valid range of statuses. We also need to ensure that
bug_resolve/bug_close logic isn't reapplied when the status changes from
one resolved status to another resolved status.