Issue #10884: Improve granularity of bugnote permissions
Currently it's only possible to use a single
$g_bugnote_allow_user_edit_delete setting to turn ON or OFF the ability
for users to be able to edit/delete their own bugnotes.

This patch provides improvement to the granularity of bugnote
permissions to make the following possible:
1) Separate edit and delete permissions
2) Thresholds instead of boolean switches
3) Ability for users to change the view state of their own bugnotes

On the topic of view states, the $g_set_view_status_threshold setting
controls whether a user can make a bug or bug note private at the time
of reporting/posting. The existing $g_change_view_status_threshold
setting allows users to change the view state of any bug or bug note
after they've been reported/posted. This patch introduces
$g_bugnote_user_change_view_state_threshold which allows administrators
to give users the the ability to make their own bugnotes public or
private as they choose, at any time.

$g_bugnote_user_edit_threshold is a new setting that can be changed to
allow users the ability to edit their own bugnotes.
$g_bugnote_user_delete_threshold is a new setting that can be changed to
allow users the ability to delete their own bugnotes. You may be
wondering why there is a difference when a user could simply edit and
blank their bugnote. The answer lies with bugnote revisions and the
ability for users to see a revision history of each bug that has been
edited.

Many thanks to Chris Fitch for the original idea and an initial patch
from August 2009. This patch is a rewrite of his earlier work with a
number of additional changes and improvements.