Issue #11404: Record dropping of bug revisions in bug history
When dropping bug revisions (bugnotes, description, steps to reproduce
and additional information) we should be recording this event in the bug
history. This maintains an audit log of who and at what time revision
data was permanently deleted.

This commit also fixes an issue whereby links in the history were still
shown for bug revisions that had been dropped. Users would thus click a
link to come across a 'revision not found' error message.