Fix #11415: Check executable bit on admin/ folder, not readable bit
login_page has a check that warns if the admin/ folder is still in
place. However this check uses is_readable() instead of is_executable().
In the case of folders, the read bit determines who can list the
directory. The execute bit determines who can traverse and access files
within the directory.

Signed-off-by: David Hicks <hickseydr@optusnet.com.au>