Fix #11417: Allow EVENT_MENU_MAIN plugin events to return null
A plugin may decide not to print a menu entry via the EVENT_MENU_MAIN
event hook based on which project is currently selected. Therefore the
return value from EVENT_MENU_MAIN should be checked to make sure it
isn't null. If it is null, then don't print anything in the menu.

Signed-off-by: David Hicks <hickseydr@optusnet.com.au>