Fix #11449: Fix incorrect table name in db_get_table call
In the 1.2.x branch of MantisBT, db_get_table expects a prefix of
"mantis" and a suffix of "table". This was updated in the 1.3.x branch
so that the prefix and suffix aren't used in calls to db_get_table.

Commit 8a9f6d0593d78bc7b4cbe47c0444ba0d50d05fa2 was a backport of the
same patch applied to the 1.3.x branch. In this backport, the
db_get_table argument should have been changed to add the required
"mantis" prefix and "table" suffix to the table name.