Issue #10059: Select 'duplicate' resolution automatically
If a bug X has a duplicate_of relationship with another bug Y and bug X
is resolved via bug_change_status_page.php, we should preselect the
resolution as 'duplicate'.

This commit also prevents the display of the duplicate bug ID field when
an existing duplicate relationship exists.

A new configuration option $g_bug_duplicate_resolution defines the
default resolution that is used when bugs are resolved as being
duplicates.