Fix #11482: Fix delimiter for links generated with print_page_link[s]
The functions print_page_link/print_page_links take a URL and add extra
parameters to the end of the URL for page navigation. The problem is
that plugin pages already contain parameters. Thus we shouldn't assume
that URLs passed to these functions don't already contain parameters.

This patch checks for existing parameters and if any are found, page
navigation parameters are appended to the end with an ampersand
delimiter instead of question mark.

Signed-off-by: David Hicks <hickseydr@optusnet.com.au>