Issue #11576: New and improved admin/check.php functionality
The existing check.php in the admin/ directory is outdated and of poor
quality to users. This check script needed a complete rewrite from
scratch to make it more user friendly, provide a wider range of checks
and have improved grammar/descriptions to make the test results easier
for users to digest.

The new check system is modular in nature so that it is easy to find and
add checks in the future as the MantisBT project grows. Thus it is
possible to include some of these "modules" in other administration
scripts (such as the installer).

Most of the old checks are carried over (with some deliberately left
out) and many new checks have been added.

This new system also includes a MantisBT installation integrity checker
that determines the hash of each file in a MantisBT installation to see
if it originates from an official release. This functionality makes it
easier for users to recognise failed upgrades (where new files haven't
overwritten old ones). More importantly, it is hoped that this
functionality can reduce the number of invalid bug reports at
mantisbt.org/bugs due to people making errors with custom "hard hacks"
to their MantisBT installations.