Fix #11518: Application error #1904 has no error description
If an invalid signup captcha is entered upon signing up a new account,
application error 1904 is raised. However, the error description
accompanying the error ID of 1904 is not shown on the error page.

This problem is due to a bug within lang_get_default() from lang_api.php
where the language of a non-authenticated user is set to boolean 'false'
instead of the default language of the entire MantisBT installation.