current_user_get_bug_filter() can return false
When calling current_user_get_bug_filter(), we need to check if the
return value is boolean false. If so, we need to get the default filter
instead of the current user's filter.

Also, current_user_get_bug_filter() already validates the filter to
ensure it is correct, so there is no need to validate it a second time.