Fix #11552: No errors shown when actiongroup tag attaching fails
When attacking tags to multiple issues via the action group feature (the
dropdown in the bottom left of the view all issues page), errors are
silently ignored. The user should instead be informed if any of the tags
could not be applied to the issues selected.

This patch also changes the return status from the actiongroup
validation and processing functions to be an error string or null in the
event that an error didn't occur. In the future it'd be much better if
the validation and processing functions worked on the entire set of
selected issues instead of each issue one-by-one. This gives the
actiongroup validation/processing functions more flexibility (especially
in the case of the tag attaching action).