Fix #11600: Bugnote direct links include mismatched parenthesis
In the bugnote direct links in the left column of the "Notes" section of
an issue page, the right parenthesis is part of the link but the left
parenthesis is not.

This patch removes the parenthesis entirely and instead uses the value
of $g_bugnote_link_tag as a prefix to the bugnote ID. This technique is
a subtle reminder to users of the syntax to use within bugnotes and bug
descriptions to refer to a particular bugnote.