Fix #11561: Filtering fails for multi-select custom field types
When attempting to filter bugs based on multi-select custom fields, some
bugs are not returned. This is due to multi-select custom field values
being stored as a string in the database with a vertical pipe character
being used as a delimiter between values. We can't just check for:

LIKE '%|value|%'

We also need to check for:

LIKE 'value|%'
LIKE '%|value'
= 'value'

To catch all possible cases of a value being stored within a
multi-select custom field string.