Fix #11624: Update version refs in child projects
When updating version names, we need to properly cascade those changes
to child projects that are using the inherited version names.  This can
potentially cause an issue with child projects that have a version with
the same name as the parent project, but I believe this mechanism better
handles the user's intent, and prevents sites from "breaking" with the
new version inheritance feature.

A better solution would be to fix versions so that they are referenced
by unique IDs everywhere, similar to how categories were upgraded to use
IDs instead of duplicating the name everywhere.  However, that solution
would not be considered within the scope of 1.2 due to schema changes
necessary for such a change.

Also, the project hierarchy was updated slightly to allow the version
API to access a full list of subprojects, including disabled projects.