Fix #11561: Database storage format of CF multilist, checkbox and radio types
Once upon a time multi-select custom field types (checkbox and
multiselect) were stored in the database in the format of
"option1|option2|option3" where they should have been stored in a format
of "|option1|option2|option3|". Additionally, radio custom field types
were being stored in the database with an unnecessary vertical pipe
prefix and suffix when there is only ever one possible value that can be
assigned to a radio field.

This patch introduces a new schema upgrade function that corrects the
invalid storage of checkbox, multiselect and radio custom fields in the
database.