Fix #0011376: New Simplified Chinese translation
Please use translatewiki.net in the future. The patch in the report was quite bad. It contained faulty headers compared to the headers used in the other translation files, as well as untranslated messages. Merging these was a bitch. Still incomplete compared to what the previous version contained. Still, updated, and complete fix will come in the next translatewiki.net export.

Signed-off-by: Siebrand Mazeland <s.mazeland@xs4all.nl>