Incorrect logic in check_selected()
The check_selected() function fails when the first argument is an array
of strings and the second argument is an integer, float or boolean
value.

This patch resolves incorrect logic that led to a === type safe
comparison between a string and another type. Obviously if we know the
two types don't match, there is no point doing a type safe comparison on
the two variables!