Issue #11729: Preselect next highest status in bug_change_status
The bug change status dropdown list currently preselects the
first/lowest status as the default. It would be more useful if the
default value selected was the next highest status from the current
status. Failing that (for closed issues) we should revert back to using
the highest available status as the default.