Fix #11530: Don't expose real name and email of higher users
manage_user_threshold grants permission to users so that they can modify
user accounts which have an equal or lower access level.

Therefore we shouldn't be exposing the real name and email of users on
view_user_page.php if the target user has a higher access level. The
manage user link should also not appear when the target user has a
higher access level.