Fix potential error loop with output plugins
In cases where a plugin hooks the page layout events, and encounters an
error that halts execution, the error API will inadvertantly trigger the
plugin's error a second time while generating the error output.  This
can potentially lead to a fatal PHP error instead of an error message.

To fix this, the event API gains a new function to clear all callbacks
hooked to all events.  This allows the error API to safely generate
output without plugin hooks being called.