Fix #11394: No email sent for lost password process
When a MantisBT installation is setup without anonymous access enabled
auth_get_current_user_id() will fail, redirecting the user back to the
login screen.

This was due to a problem in user_pref_get() where it was assumed that a
user was logged in. For the lost password recovery step (with no
anonymous login), the lost password recovery step happens without any
user logged in.