Fix #11817: Modified MENU_ISSUE for consistency
In order to be more consistent with other menu events, EVENT_MENU_ISSUE
has been modified to allow both the existing method of adding links as
well as the ability to add links the same way as for other menu events.

The existing behavior is triggered by an array with non-numeric keys
defining the link's label, while the new behavior is triggered by
numeric keys, or by not returning a single string instead of an array.

TODO: Update the other menu events to use this combined mechanism for
ultimate consistency.