Fix #11641: Update history table on version rename
When a version is renamed, existing bug history entries should be
updated with the new version name, including history entries for bugs
from subprojects in the case where versions are inherited from parents.

This has the same potential edge case as previous commits in this area,
such that any child project that contains a duplicate version name will
have entries updated in that child project even if that's not the
desired behavior.  However, this is the best option until versions can
be migrated to using integer IDs the same as categories were.