Make feedback reassignment optional
In certain installations, the new behavior of automatically setting an
issue back to new or assigned status is unwanted and disruptive to their
existing workflow.  In these cases, there should be a way to disable
this new feature so that the issue will stay in feedback status until
explicitly changed.

The new configuration option is "reassign_on_feedback"', and the default
value for new installations is ON.