Fix #11839: Get Time Tracking Information button was removed
The "Get Time Tracking Information" button was removed as part of issue
view bug page and allows users to get accumulated totals of time spent
by each developer on a single issue. This is useful for very large
issues with hundreds of comments.

Thanks to Doug Campbell for reporting this issue and providing an
initial patch.