Fix #11862: Preselect user's current access level in dropdown
On manage_user_edit_page.php the access level of a user is always
preselected to be the default access level at which new user accounts
are created. We should instead try to preselect the user's current
access level from the list so that when the form is saved, the access
level of the user isn't changed by accident.

Also remove the fake [Default Access Level] option from the list as it
serves little to no useful purpose.