Issue #11896: Remove [^] "open in new window" suffix from links
Modern browsers make it easy for users to open links in a new window via
a context menu or shortcut key on the keyboard. Thus this feature is
more-or-less redundant. It was also not clear to users that the
characters [^] denote that a link will open in a new window.