Issue #11897: Refactor footer of pages using XHTML/CSS
The footer of each page is currently a bit of a mess using an old school
table based layout.

The following changes help improve the footer:

1) "Copyright MantisBT group" makes it sound like the content of the bug
tracker is copyrighted by MantisBT when in fact it is just the software
which is copyrighted. This should be changed to "Powered by MantisBT
which is copyrighted by..." to remove the ambiguity.

2) Introduce a new (optional) $g_copyright_statement configuration
option that allows users to specify in a single line the copyright and
licensing status of the content of their bug trackers.

3) Add a new $g_show_memory_usage setting so that $g_show_time doesn't
pull in the memory usage as well.

4) $g_show_queries_count should be properly separated from
$g_show_queries_list. The count is just a number and doesn't cause any
harm. The queries list could contain potentially sensitive information
and thus is treated as a completely separate option from showing just
the query count.

5) Use semantic XHTML markup for the footer with CSS styling (instead of
the old table based layout).