Fix #11594: No column title printed for attachment_count
When $g_show_attachment_indicator = ON, no column title was printed on
view_all_bug_page.php for the attachment_count column. This led to
formatting issues with the table where column titles were aligned to the
wrong columns.

This patch also improves the accessibility of the attachment icons and
links using alt/title tags to provide the user with more information on
the attachment links.