Only allow NuSOAP to compress SOAP responses
When NuSOAP receives a request which declares that it accepts a compressed (gzip,deflate)
response, it performs its own compression. Although conceptually incorrect, since this
should be handled by the web server, there is no configuration flag to disable it.

Therefore we disable all forms of compression for the SOAP API and allow NuSOAP to perform
the compression itself.

Fixes #11868: returned compressed data corrupt