install_stored_filter_migrate should ignore missing filter fields
Commit c42ac127bd05851d62a9ab912229d8c279f760c9 standardised field names
between the database and filters subsystem of MantisBT. However the
migration function (install_stored_filter_migrate) expected old filter
strings to contain a value for every filterable field. When this
condition was not met (a filter string didn't contain any reference to
the show_category field for example), the migration function failed with
an error.