Fix edge case with event_clear_callbacks
In cases where an error would occur before Mantis had fully
bootstrapped, it was possible for the error API to generate misleading
errors when the event API had not yet been loaded.  By wrapping the call
to event_clear_callbacks() with a check for function_exists(), it should
fix this sort of behavior.