Issue #12096: Add reporter and handler to monitor list on duplicate
When an issue X is resolved as a duplicate of an existing issue Y, the
monitor list from X is copied to and merged with Y. However the reporter
and handler of the source issue are not copied unless they're on the
monitor list.

This change ensures that the reporter and handler of the source issue
are added to the monitor list on the destination issue. These users are
thus less likely to become out-of-the-loop on an issue that is relevant
to them.