Added form_action_self() API function
Helper function to generate a form action value when forms are designed
to be submitted to the same url that's is currently being used, such as
helper_ensure_confirmed() or auth_reauthenticate().

Designed to fix issues with using reverse proxies that rewrite URLs for
the application server.  By using a standard self-relative form action,
it prevents issues where SCRIPT_NAME on the application server does not
match the URL that the client is using.