Fix #11967: Fix bug update events (both pre and post commit)
The bug update events have been modified to be more useful to plugin
developers.

EVENT_UPDATE_BUG_DATA is a chained event type which allows plugins to
compare the existing and updated bug data objects and make changes to
either of these objects as appropriate. This event can also be used to
invalidate a bug update for not meeting more complex rules.

EVENT_UPDATE_BUG has been modified to execute after the update is
committed to the database and prior to email and other notifications
being sent. Again the original and updated bug data objects are provided
so that plugins can compare the two objects to get a list of fields that
changed during the update.