Fix #11981: Factor in view status when showing user monitor list
When a user attempts to send a reminder to users about an issue they're
presented with a list of users from which they can select from. This
list of users is generated by determining the minimum access level
required to view the issue. However the view state of the issue was not
taken into consideration and thus if an issue had a private view state,
users without permission to view private issues would still be shown in
the user list.

Thanks to Sergio Del Franco for reporting this issue and providing a
patch.