Fix #11852: Forgotton password feature prevents admin password reset
When the forgotton password feature is turned off in config_inc.php this
removes the [Forgotton Password] link from the login page.
It also has the unintented side effect of preventing e-mails being sent
out for a password reset through the Management console.
This is caused by the use of 'lost_password_feature' in an if clause in
email_api.php.

Thanks to Alex Russell (xavierblue) for this patch.