Issue #11826: Implement new loading functions for CSS and JS
Two new functions have been added so that MantisBT pages can inform the
MantisBT core (html_api.php in particular) which JavaScript files and
stylesheets to load during the output HTML rendering.

These functions must be used PRIOR to html_top() being executed as after
that point of time it'll be too late to load new scripts or stylesheets.

This functionality will prevent multiple inclusions of the same
JavaScript file and will also prevent the need for downloading a new
stylesheet or script half way through the rendering of the page within
the user's browser.